% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMAPPR2data.R
\docType{package}
\name{MMAPPR2data}
\alias{MMAPPR2data}
\alias{zy13wt}
\alias{zy13wtIdx}
\alias{zy13mut}
\alias{zy13mutIdx}
\alias{downloadAll}
\alias{MMAPPR2data-package}
\title{MMAPPR2data: Sample Data for MMAPPR2}
\description{
Contains BAM files and indices for example use in MMAPPR2.
The \emph{zy13} mutation was identified and mapped using RNA-Seq as
described in Hill et al. in Genome Research (2013).
The \code{.fastq} files for the two pools were
downloaded from the B2B Consortium's GNomEx datahub and were aligned to
the GRCz11 genome using HISAT2 and being cut and filtered to include
only reads on chromosome 7 with high mapping quality.
See \code{scripts/make-data.sh} for details.
}
\details{
Abbreviations used in resource names:
wt=wild-type pool, mut=mutant pool, Idx=BAM index.
}
\examples{
library(ExperimentHub)

eh <- ExperimentHub()
wtfiles <- listResources(eh, "MMAPPR2data", "wt")
wtfiles[[1]]        ## load the first resource in the list
## load all mutant files
mutfiles <- loadResources(eh, "MMAPPR2data", "mut")

## Files can also be accessed directly like this:
zy13wt() ## data are loaded
zy13wt(metadata = TRUE)  ## metadata are displayed

## Download all resources at once:
downloadAll()

}
