% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zy13candidates}
\alias{zy13candidates}
\title{Candidates list for zy13 example}
\format{A list containing only one element: a
\code{\link[GenomicRanges]{GRanges}} object containing
ranked candidate mutations resulting from the \code{generateCandidates}
step of the MMAPPR2 pipeline for chromosome 7.}
\usage{
zy13candidates
}
\description{
A list representing the candidates info resulting from running
the MMAPPR2 pipeline on the sample data contained in this
package.
}
\details{
The typical user should not have to concern themselves with this
dataset; it is used to enable dynamic examples in the development of
the MMAPPR2 package.
}
\keyword{datasets}
