% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMAPPR2data.R
\docType{package}
\name{MMAPPR2data}
\alias{MMAPPR2data}
\alias{exampleMutBam}
\alias{exampleWTbam}
\alias{MMAPPR2data-package}
\title{MMAPPR2data: Example Data for MMAPPR2}
\usage{
exampleMutBam()

exampleWTbam()
}
\value{
A \code{\link[Rsamtools:BamFile-class]{BamFile}} object referencing
  a BAM file and its index.

A \code{\link[Rsamtools:BamFile-class]{BamFile}} object referencing
  a BAM file and its index.
}
\description{
Contains BAM files and indices for example use in MMAPPR2.
The data is artificial, meant to simulate sequencing of the zebrafish slc24a5
gene in mutant and wild-type pools resulting from the cross of a novel
mutant from a forward genetics screen with a wild-type line, as described in
Hill et al. 2013.
}
\section{Functions}{
\itemize{
\item \code{exampleMutBam}: Easy access to example mutant pool BAM file.

\item \code{exampleWTbam}: Easy access to example wild-type pool BAM file.
}}

\examples{
exampleMutBam()
exampleWTbam()

mutFile <- exampleMutBam()
wtFile <- exampleWTbam()
}
